simpliying menu lookup
